mv feature and feature namespace info
